Partial refactoring
