+++ .gitignore
@@ -24,7 +24,6 @@ tmpout
 ## Specific to RubyMotion:
 .dat*
 .repl_history
-.build.lockfile
 build/
 *.bridgesupport
 build-iPhoneOS/
