+++ src/modules/common_strings.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+def exit_string
+  'Press Enter to exit.'.cyan
+end
+
+def separator
+  '____________________________________________________________'.light_black
+end
