+++ src/modules/handlers.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require_relative 'handlers/generic'
+
+require_relative 'handlers/champions'
+require_relative 'handlers/emotes'
+require_relative 'handlers/eternals'
+require_relative 'handlers/exception'
+require_relative 'handlers/icons'
+require_relative 'handlers/materials'
+require_relative 'handlers/skins'
+require_relative 'handlers/wards'
