+++ src/modules/handlers/emotes.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+def handle_emotes
+  handle_generic('Emotes', 'EMOTE', 'EMOTE_disenchant')
+end
