+++ src/modules/handlers/eternals.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+def handle_eternals
+  handle_generic(
+    'Eternal Shards',
+    'STATSTONE_SHARD',
+    'STATSTONE_SHARD_DISENCHANT'
+  )
+  handle_generic('Eternals', 'STATSTONE', 'STATSTONE_DISENCHANT')
+end
