+++ src/modules/handlers/icons.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+def handle_icons
+  handle_generic('Icons', 'SUMMONERICON', 'SUMMONERICON_disenchant')
+end
