+++ src/modules/handlers/skins.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+def handle_skins
+  handle_generic('Skin Shards', 'SKIN_RENTAL', 'SKIN_RENTAL_disenchant')
+  handle_generic('Skin Permanents', 'SKIN', 'SKIN_disenchant')
+end
